Create Nodo.java
This file is the Constructor of the object called "Nodo", which is the main component of a network. This part of the code was principally made by Dr. Maximino Aldana.
This Nodo (or node) has several parameters which identificate it:

name
Integer state
double state
relaxation coefficient
initial state
activation threshold value
number of regulators
regulatory function

with these parameters one can determine the dynamics of each node and in turn, the dynamical state of the whole network