Create Attractor.java
This is the constructor of the object called Attractor. This is, the steady state of a network, a set of configurations of the nodes which repeat after certain time. With this file it is possible to know all the attractors of a network, the period of the attractor, their basins of attraction (the number of initial conditions that reach the same attractor) and the dynamical states of each node inside the attractor.